Fixed mistakes with Skeletal Mesh vao

+++ assets/shaders/skeletalMesh.vert
@@ -2,7 +2,7 @@
 layout (location = 0) in vec3 vertexPosition;
 layout (location = 1) in vec3 vertexNormal;
 layout (location = 2) in vec2 vertexTexCoord;
-layout (location = 3) in vec4 vertexJoints;
+layout (location = 3) in ivec4 vertexJoints;
 layout (location = 4) in vec4 vertexWeights;
 
 out vec3 fragPosition;
@@ -18,10 +18,10 @@ uniform mat4[64] matJoints;
 void main()
 {
     mat4 matBonesCombined = 
-        matJoints[int(vertexJoints.x)] * vertexWeights.x +
-        matJoints[int(vertexJoints.y)] * vertexWeights.y +
-        matJoints[int(vertexJoints.z)] * vertexWeights.z +
-        matJoints[int(vertexJoints.w)] * vertexWeights.w;
+        matJoints[vertexJoints.x] * vertexWeights.x +
+        matJoints[vertexJoints.y] * vertexWeights.y +
+        matJoints[vertexJoints.z] * vertexWeights.z +
+        matJoints[vertexJoints.w] * vertexWeights.w;
 
     mat4 matSkin = matModel * matBonesCombined;
     vec4 worldPos = matSkin * vec4(vertexPosition, 1.0);
